# openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW BFF API
  description: |
    Backend for Frontend (BFF) REST API que agrega e simplifica o acesso aos microserviços do sistema ConstrSW.
    
    Esta API atua como uma camada intermediária entre o frontend e os microserviços, fornecendo:
    - Agregação de dados de múltiplos serviços
    - Simplificação de chamadas complexas
    - Redução do número de requisições necessárias
    - Padronização de respostas
    
    ## Autenticação
    Todos os endpoints (exceto `/auth/login` e `/auth/refresh`) requerem autenticação via Bearer Token (JWT).
    O token deve ser obtido através do endpoint `/auth/login` e incluído no header `Authorization: Bearer <token>`.
    
    ## Microserviços Agregados
    - OAuth (Autenticação e Autorização)
    - Employees (Funcionários)
    - Classes (Turmas)
    - Courses (Disciplinas)
    - Lessons (Aulas)
    - Professors (Professores)
    - Reservations (Reservas)
    - Resources (Recursos)
    - Rooms (Salas)
    - Students (Estudantes)
  version: 1.0.0
  contact:
    name: ConstrSW Team
    email: support@constrsw.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desenvolvimento local
  - url: http://api:3000/api/v1
    description: Servidor interno (Docker)
...
paths:
...
  /users/{userId}:
    patch:
      tags:
        - Users
      summary: Atualizar senha de um usuário
      description: Atualiza a senha de um usuário
      operationId: updateUserPassword
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Senha atualizada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /reservations/{reservationId}/authorized-users:
    get:
      tags:
        - Reservations
      summary: Listar usuários autorizados de uma reserva
      operationId: listReservationAuthorizedUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        '200':
          description: Lista de usuários autorizados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorizedUserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Reservations
      summary: Adicionar usuário autorizado à reserva
      operationId: createReservationAuthorizedUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizedUserCreateRequest'
      responses:
        '201':
          description: Usuário autorizado adicionado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedUserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /reservations/{reservationId}/authorized-users/{authorizedUserId}:
    get:
      tags:
        - Reservations
      summary: Buscar usuário autorizado específico
      operationId: getReservationAuthorizedUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/AuthorizedUserId'
      responses:
        '200':
          description: Dados do usuário autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedUserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Reservations
      summary: Atualizar usuário autorizado (completo)
      operationId: updateReservationAuthorizedUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/AuthorizedUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizedUserUpdateRequest'
      responses:
        '200':
          description: Usuário autorizado atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedUserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Reservations
      summary: Atualizar usuário autorizado (parcial)
      operationId: patchReservationAuthorizedUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/AuthorizedUserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizedUserUpdateRequest'
      responses:
        '200':
          description: Usuário autorizado atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedUserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Reservations
      summary: Remover usuário autorizado da reserva
      operationId: deleteReservationAuthorizedUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReservationId'
        - $ref: '#/components/parameters/AuthorizedUserId'
      responses:
        '204':
          description: Usuário autorizado removido com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /categories:
    get:
      tags:
        - Resources
      summary: Listar todas as categorias
      operationId: listCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Resources
      summary: Criar uma nova categoria
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Categoria criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /categories/{categoryId}:
    get:
      tags:
        - Resources
      summary: Buscar categoria por ID
      operationId: getCategoryById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Dados da categoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Resources
      summary: Atualizar categoria (parcial)
      operationId: patchCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Categoria atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Resources
      summary: Substituir categoria (completo)
      operationId: replaceCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Categoria substituída com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Resources
      summary: Deletar categoria
      operationId: deleteCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Categoria deletada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /feature-values:
    get:
      tags:
        - Resources
      summary: Listar todos os feature values
      operationId: listFeatureValues
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Resources
      summary: Criar um novo feature value
      operationId: createFeatureValue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueCreateRequest'
      responses:
        '201':
          description: Feature value criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /feature-values/resource/{resourceId}:
    get:
      tags:
        - Resources
      summary: Buscar feature values por recurso
      operationId: getFeatureValuesByResource
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /feature-values/feature/{featureId}:
    get:
      tags:
        - Resources
      summary: Buscar feature values por feature
      operationId: getFeatureValuesByFeature
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /feature-values/resources/{resourceId}/features:
    get:
      tags:
        - Resources
      summary: Listar feature values de um recurso
      operationId: listResourceFeatureValues
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Resources
      summary: Criar feature value para um recurso
      operationId: createResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueCreateRequest'
      responses:
        '201':
          description: Feature value criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /feature-values/resources/{resourceId}/features/{featureValueId}:
    get:
      tags:
        - Resources
      summary: Buscar feature value específico
      operationId: getResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      responses:
        '200':
          description: Dados do feature value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Resources
      summary: Atualizar feature value
      operationId: patchResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueUpdateRequest'
      responses:
        '200':
          description: Feature value atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Resources
      summary: Deletar feature value
      operationId: deleteResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      responses:
        '204':
          description: Feature value deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /auth/login
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: ID único do usuário
      schema:
        type: string
...
    ReservationId:
      name: reservationId
      in: path
      required: true
      description: ID único da reserva
      schema:
        type: string
    ResourceId:
      name: resourceId
      in: path
      required: true
      description: ID único do recurso (UUID)
      schema:
        type: string
        format: uuid
...
    AuthorizedUserId:
      name: authorizedUserId
      in: path
      required: true
      description: ID único do usuário autorizado
      schema:
        type: string
        format: uuid
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: ID único da categoria
      schema:
        type: string
        format: uuid
    FeatureId:
      name: featureId
      in: path
      required: true
      description: ID único da feature
      schema:
        type: string
    FeatureValueId:
      name: featureValueId
      in: path
      required: true
      description: ID único do feature value
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Mensagem de erro
        status_code:
          type: integer
          description: Código HTTP do erro
...
    AuthorizedUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    AuthorizedUserCreateRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    AuthorizedUserUpdateRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    CategoryCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    CategoryUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    FeatureValueResponse:
      type: object
      properties:
        id:
          type: string
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid
    FeatureValueCreateRequest:
      type: object
      required:
        - value
        - featureId
        - resourceId
      properties:
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid
    FeatureValueUpdateRequest:
      type: object
      properties:
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autorizado - Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'